+++ src/modules/addWMS/store/stateAddWMS.js
@@ -7,6 +7,7 @@
  * @property {String[]} supportedDevices Devices on which the module is displayed.
  * @property {String[]} supportedMapModes Map mode in which this module can be used.
  * @property {String[]} exampleURLs List of example WMS service URLs that will be shown to the user.
+ * @property {Number} [featureCount=undefined] Amount of features retrieved on GFI requests. Corresponds to the GetFeatureInfo parameter "FEATURE_COUNT". If undefined is specified, the attribute FEATURE_COUNT will not be listed in the gfi request.
  */
 const state = {
     description: "common:modules.addWMS.description",
@@ -15,7 +16,8 @@ const state = {
     icon: "bi-cloud-plus",
     supportedDevices: ["Desktop", "Mobile", "Table"],
     supportedMapModes: ["2D", "3D"],
-    exampleURLs: []
+    exampleURLs: [],
+    featureCount: undefined
 };
 
 export default state;